# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateOther:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      mavzular:
        type: string
      middle_name:
        type: string
      oqish_joyi:
        type: string
      password:
        type: string
      phone_number:
        type: string
      yonalish:
        type: string
    type: object
...
  models.UpdateOther:
    properties:
      Yonalish:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      mavzular:
        type: string
      middle_name:
        type: string
      oqish_joyi:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /other:
    get:
      consumes:
      - application/json
      description: Get List Other
      operationId: get_list_other
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Other
      tags:
      - Other
    post:
      consumes:
      - application/json
      description: Create Other
      operationId: create_other
      parameters:
      - description: CreateOtherRequest
        in: body
        name: other
        required: true
        schema:
          $ref: '#/definitions/models.CreateOther'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Other
      tags:
      - Other
  /other/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Other
      operationId: delete_other
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Other
      tags:
      - Other
    get:
      consumes:
      - application/json
      description: Get By ID Other
      operationId: get_by_id_other
      parameters:
      - description: id
        in: path
        name: id
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get By ID Other
      tags:
      - Other
    put:
      consumes:
      - application/json
      description: Update Other
      operationId: update_other
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateOtherRequest
        in: body
        name: other
        required: true
        schema:
          $ref: '#/definitions/models.UpdateOther'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Other
      tags:
      - Other
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"